Get rid of the default_set_options attribute, since it's really just a computer value.
Also got rid of the BUILD by adding a separate builder method for
determining the driver_short_name.

Added a Duration type for expires_in, and added a coercion to coerce
it from a string. This required a change to one test, since
expires_at() now always returns an integer, not the value originally
passed to it. This is backwards incompatible, but I can't imagine
anyone relies on this.

Jon, what do you think?


git-svn-id: https://perl-cache.googlecode.com/svn/chi/trunk@454 fe4db14e-1138-0410-ae3a-bd28db37d31a