Replace List::MoreUtils with List::Util
List::Util was already used (for shuffle & sum) but depend on 1.45
as that's when uniq was added.

List::Util is dual-life so from Perl 5.26 the version we need will
already be there, saving the user an install.

Some of the uses of List::{MoreUtils,Util} were actually unused so just
drop them.